Updated commons text and some esthetical improvements
